Refactor Segment Library
