+++ eidolon/logger.py
@@ -5,7 +5,6 @@ import logging
 
 from eidolon.environment import APPLICATION_NAME, LOG_LEVEL
 
-# Initialize logger for the application
 logger = logging.getLogger(APPLICATION_NAME)
 logger.setLevel(LOG_LEVEL)
 
@@ -18,13 +17,11 @@ def log_lambda_statistics(event, context) -> None:
         event: The event that triggered the Lambda function.
         context: The context in which the Lambda function is running.
     """
-    # Log basic Lambda function details
     if context:
         logger.info(f"Function: {context.function_name}")
         logger.debug(f"Memory: {context.memory_limit_in_mb}")
         logger.debug(f"Time Remaining: {context.get_remaining_time_in_millis()}")
 
-    # Log details from the event, particularly for authentication and authorization
     if event:
         claims = event.get("requestContext", {}).get("authorizer", {}).get("claims", {})
         logger.info(f"User: {claims.get('cognito:username')}")
