+++ scripts/test_fixes.py
@@ -11,7 +11,8 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
 from eidolon.schema import normalize_segment_definition
-from eidolon.segment import process_skill_challenges, validate_segment_outcome_results
+from eidolon.segment_challenges import process_skill_challenges
+from eidolon.segment_core import validate_segment_outcome_results
 
 
 def load_test_story():
